pyproject.toml
[tool.poetry]
name = "zzzero-agent"
version = "0.1.0"
description = "基于节点编排的AI Agent框架"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.0"
asyncio = "^3.4.3"
aiohttp = "^3.9.0"
gradio = "^4.0.0"
mcp = "^0.1.0"
jinja2 = "^3.1.2"
networkx = "^3.2"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
sqlalchemy = "^2.0.0"
rich = "^13.0.0"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
ipython = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 